fix: AmountPercentAverage 오차 수정 하는 로직 반영 및 Average percent long 으로 변경, 테스트 코드 작성

+++ budget-application/src/main/java/com/hg/budget/application/budget/infrastructure/average/Average.java
@@ -2,9 +2,9 @@ package com.hg.budget.application.budget.infrastructure.average;
 
 import com.hg.budget.domain.category.Category;
 
-public record Average(Category category, double percent) {
+public record Average(Category category, long percent) {
 
-    public Average plus(double percent) {
+    public Average plus(long percent) {
         return new Average(category, this.percent + percent);
     }
 }
